Pantries: create, add, edit,delete product

+++ src/main/java/ca/foc/dao/PantryRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PantryRepository extends CrudRepository<Pantry, Integer> {
+public interface PantryRepository extends CrudRepository<Pantry, Long> {
 
 }
